% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{add_quadrant_label_nudges}
\alias{add_quadrant_label_nudges}
\title{Bias repelled labels toward Q2/Q4 (sparse for signed correlations)}
\usage{
add_quadrant_label_nudges(label_df)
}
\description{
Bias repelled labels toward Q2/Q4 (sparse for signed correlations)
}
\keyword{internal}
