% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginals.R
\name{apply_prior_ecdf}
\alias{apply_prior_ecdf}
\title{Apply prior empirical CDF to a vector}
\usage{
apply_prior_ecdf(x, knots, n_ref)
}
\description{
Apply prior empirical CDF to a vector
}
\keyword{internal}
