% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_update.R
\name{build_prior_pseudo_fit}
\alias{build_prior_pseudo_fit}
\title{Build a pseudo graphical fit from an implied correlation matrix}
\usage{
build_prior_pseudo_fit(impliedCor, nObs = NA_integer_)
}
\arguments{
\item{impliedCor}{Symmetric correlation matrix.}

\item{nObs}{Reference sample size for metadata.}
}
\description{
Build a pseudo graphical fit from an implied correlation matrix
}
\keyword{internal}
