% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{copula_ggsave}
\alias{copula_ggsave}
\title{Save a ggplot as PNG and PDF with white background}
\usage{
copula_ggsave(path, plot, width, height, dpi = 150)
}
\description{
Save a ggplot as PNG and PDF with white background
}
\keyword{internal}
