% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{extract_pairwise_values}
\alias{extract_pairwise_values}
\title{Extract all pairwise values from a symmetric matrix}
\usage{
extract_pairwise_values(
  mat,
  vars = colnames(mat),
  edge_only = FALSE,
  adj = NULL
)
}
\description{
Extract all pairwise values from a symmetric matrix
}
\keyword{internal}
