% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{ExtractEdges}
\alias{ExtractEdges}
\title{Extract edge list from a copula fit result}
\usage{
ExtractEdges(result, label)
}
\arguments{
\item{result}{Output of \code{\link[=FitStratumCopula]{FitStratumCopula()}}.}

\item{label}{Stratum label.}
}
\value{
Data frame with columns \code{from}, \code{to}, \code{pcor}, \code{stratum}, \code{edge}.
}
\description{
Extract edge list from a copula fit result
}
