% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{format_comparison_edge_label}
\alias{format_comparison_edge_label}
\title{Shorten edge names for comparison plot labels}
\usage{
format_comparison_edge_label(edges)
}
\description{
Shorten edge names for comparison plot labels
}
\keyword{internal}
