% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{format_node_labels}
\alias{format_node_labels}
\title{Format variable names for plot labels}
\usage{
format_node_labels(vars)
}
\description{
Format variable names for plot labels
}
\keyword{internal}
