% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_bayes.R
\name{get_factor_update_stan_model}
\alias{get_factor_update_stan_model}
\title{Compile or retrieve cached Stan model for factor update}
\usage{
get_factor_update_stan_model()
}
\value{
A \code{CmdStanModel} object.
}
\description{
Compile or retrieve cached Stan model for factor update
}
\keyword{internal}
