% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_test.R
\name{pairwise_tau_stat}
\alias{pairwise_tau_stat}
\title{Pairwise Kendall tau deviation from implied prior correlations}
\usage{
pairwise_tau_stat(uniformMatrix, impliedCor)
}
\arguments{
\item{uniformMatrix}{Matrix of uniform pseudo-observations (n x d).}

\item{impliedCor}{Prior implied correlation matrix.}
}
\value{
Mean absolute deviation of sample taus from implied taus.
}
\description{
Pairwise Kendall tau deviation from implied prior correlations
}
\keyword{internal}
