% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{PlotCopulaCorHeatmap}
\alias{PlotCopulaCorHeatmap}
\title{Plot copula correlation heatmap}
\usage{
PlotCopulaCorHeatmap(result, title = "Copula Correlation Matrix", vars = NULL)
}
\arguments{
\item{result}{Output of \code{\link[=FitStratumCopula]{FitStratumCopula()}}.}

\item{title}{Plot title.}

\item{vars}{Optional subset of variables to display.}
}
\value{
pheatmap grob.
}
\description{
Plot copula correlation heatmap
}
