% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_mixed.R
\name{PlotLatentMixedResidualNetwork}
\alias{PlotLatentMixedResidualNetwork}
\title{Plot residual graphical network from a latent mixed fit}
\usage{
PlotLatentMixedResidualNetwork(
  fit,
  title = "Residual copula network",
  printPlot = TRUE,
  ...
)
}
\arguments{
\item{fit}{Output of \code{\link[=FitCopulaLatentMixedModel]{FitCopulaLatentMixedModel()}}.}

\item{title}{Plot title.}

\item{printPlot}{Passed to \code{\link[=PlotCopulaNetwork]{PlotCopulaNetwork()}}.}

\item{...}{Passed to \code{\link[=PlotCopulaNetwork]{PlotCopulaNetwork()}}.}
}
\value{
ggraph / ggplot object.
}
\description{
Plot residual graphical network from a latent mixed fit
}
