% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_mixed.R
\name{PlotLatentMixedResidualPcorHeatmap}
\alias{PlotLatentMixedResidualPcorHeatmap}
\title{Plot residual partial-correlation heatmap from a latent mixed fit}
\usage{
PlotLatentMixedResidualPcorHeatmap(
  fit,
  title = "Residual partial correlation",
  ...
)
}
\arguments{
\item{fit}{Output of \code{\link[=FitCopulaLatentMixedModel]{FitCopulaLatentMixedModel()}}.}

\item{title}{Plot title.}

\item{...}{Passed to \code{\link[=PlotCopulaCorHeatmap]{PlotCopulaCorHeatmap()}}.}
}
\value{
pheatmap object.
}
\description{
Plot residual partial-correlation heatmap from a latent mixed fit
}
