% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{PlotPcorHeatmap}
\alias{PlotPcorHeatmap}
\title{Plot partial correlation heatmap}
\usage{
PlotPcorHeatmap(
  result,
  title = "Partial Correlation Matrix (Glasso)",
  vars = NULL,
  zeroDiag = TRUE
)
}
\arguments{
\item{result}{Output of \code{\link[=FitStratumCopula]{FitStratumCopula()}}.}

\item{title}{Plot title.}

\item{vars}{Optional subset of variables to display.}

\item{zeroDiag}{If \code{TRUE}, zero the diagonal for display.}
}
\value{
pheatmap grob.
}
\description{
Plot partial correlation heatmap
}
