% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{resolve_dim}
\alias{resolve_dim}
\title{Resolve save dimension with optional override}
\usage{
resolve_dim(default, override = NULL)
}
\description{
Resolve save dimension with optional override
}
\keyword{internal}
