% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{SaveCheckpoint}
\alias{SaveCheckpoint}
\title{Save a checkpoint RDS file}
\usage{
SaveCheckpoint(object, outDir, filename = "checkpoint.rds")
}
\arguments{
\item{object}{R object to save.}

\item{outDir}{Output directory.}

\item{filename}{File name (default \code{checkpoint.rds}).}
}
\value{
Invisibly returns the full file path.
}
\description{
Save a checkpoint RDS file
}
