% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_prior.R
\name{SimulateFactorCopula}
\alias{SimulateFactorCopula}
\title{Simulate data from a fitted factor copula prior}
\usage{
SimulateFactorCopula(priorFit, nSim = NULL, nObs = NULL)
}
\arguments{
\item{priorFit}{Output of \code{\link[=FitFactorCopulaPrior]{FitFactorCopulaPrior()}}.}

\item{nSim}{Number of simulated observations.}

\item{nObs}{Alias for \code{nSim} (either may be used).}
}
\value{
Numeric matrix (nSim x d) of simulated continuous scores on the
normal-copula scale.
}
\description{
Simulate data from a fitted factor copula prior
}
