% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{symmetric_matrix_breaks}
\alias{symmetric_matrix_breaks}
\title{Symmetric color breaks centered at 0, bounded by max |value| in mat}
\usage{
symmetric_matrix_breaks(mat, n = 101)
}
\description{
Symmetric color breaks centered at 0, bounded by max |value| in mat
}
\keyword{internal}
